requirements.txt
flask>=3.0.0
requests>=2.31.0
click>=8.1.0
Pillow>=10.0.0
aiohttp>=3.9.0
pystray>=0.19.5
